fixing Satellite6 ISO installer to make node-installer too. deltacloud.keepalive announced

+++ src/com/redhat/qe/katello/tests/e2e/SetupServers.java
@@ -44,7 +44,8 @@ public class SetupServers extends KatelloCliTestBase {
 	
 	@AfterSuite(alwaysRun=true)
 	public void tearDown() {
-		if (isDeltacloud) {
+		boolean keepAlive = Boolean.getBoolean(System.getProperty("deltacloud.keepalive", "false"));
+		if (isDeltacloud && !keepAlive) {
 			KatelloUtils.destroyDeltaCloudMachine(server);
 			for (DeltaCloudInstance client : clients) {
 				KatelloUtils.destroyDeltaCloudMachine(client);
