+++ src/com/redhat/qe/katello/tests/upgrade/v1/SetupServers.java
@@ -46,12 +46,12 @@ public class SetupServers extends KatelloCliTestBase {
 	
 	@AfterSuite(alwaysRun=true)
 	public void tearDown() {
-		if (isDeltacloud) {
+		boolean keepAlive = Boolean.getBoolean(System.getProperty("deltacloud.keepalive", "false"));
+		if (isDeltacloud && !keepAlive) {
 			KatelloUtils.destroyDeltaCloudMachine(server);
 			KatelloUtils.destroyDeltaCloudMachine(client);
 			KatelloUtils.destroyDeltaCloudMachine(client2);
 			KatelloUtils.destroyDeltaCloudMachine(client3);
 		}
 	}
-
 }
